# docs/openapi.yaml
openapi: 3.0.1
info:
  title: Bank Cards API
  description: REST API для управления банковскими картами
  contact:
    name: Bank Cards Team
    email: support@bankcards.com
  license:
    name: Apache 2.0
    url: http://springdoc.org
  version: 1.0.0
servers:
  - url: http://localhost:8080/api
    description: Generated server url
security:
  - bearerAuth: []
...
paths:
  /admin/users/{userId}/role:
    put:
      tags:
        - Admin Controller
      summary: Изменить роль пользователя
      description: Изменяет роль пользователя (USER/ADMIN). Только для ADMIN
      operationId: updateUserRole
      parameters:
        - name: userId
          in: path
          description: ID пользователя
          required: true
          schema:
            type: integer
            format: int64
          example: 1
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateRoleRequest'
        required: true
      responses:
        '200':
          description: Роль успешно изменена
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/UserResponse'
        '400':
          description: 'Неверная роль (допустимо: USER, ADMIN)'
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Не авторизован
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Доступ запрещен
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Пользователь не найден
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /admin/cards/{cardId}/status:
    put:
      tags:
        - Admin Controller
      summary: Изменить статус карты
      description: Блокирует или активирует карту. Только для ADMIN
      operationId: updateCardStatus
      parameters:
        - name: cardId
          in: path
          description: ID карты
          required: true
          schema:
            type: integer
            format: int64
          example: 1
        - name: status
          in: query
          description: Новый статус
          required: true
          schema:
            type: string
            enum:
              - ACTIVE
              - BLOCKED
              - EXPIRED
          example: BLOCKED
      responses:
        '200':
          description: Статус карты успешно изменен
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/CardResponse'
        '400':
          description: 'Неверный статус (допустимо: ACTIVE, BLOCKED)'
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Не авторизован
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Доступ запрещен
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Карта не найдена
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /admin/users/{userId}/cards:
    get:
      tags:
        - Admin Controller
      summary: Получить карты пользователя
      description: >-
        Возвращает список карт указанного пользователя с пагинацией и поиском.
        Только для ADMIN
      operationId: getUserCards
      parameters:
        - name: userId
          in: path
          description: ID пользователя
          required: true
          schema:
            type: integer
            format: int64
          example: 1
        - name: search
          in: query
          description: 'Поиск по маске номера карты (например: 1234)'
          required: false
          schema:
            type: string
        - name: pageable
          in: query
          description: 'Пагинация: page, size, sort (например: sort=createdAt,desc)'
          required: true
          schema:
            $ref: '#/components/schemas/Pageable'
      responses:
        '200':
          description: Список карт успешно получен
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/PageCardResponse'
        '400':
          description: Неверный ID пользователя или параметры запроса
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Не авторизован
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Доступ запрещен - требуется роль ADMIN
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Пользователь не найден
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    post:
      tags:
        - Admin Controller
      summary: Создать карту для пользователя
      description: >-
        Создает новую карту для указанного пользователя. Номер генерируется
        автоматически. Только для ADMIN
      operationId: createCardForUser
      parameters:
        - name: userId
          in: path
          description: ID пользователя
          required: true
          schema:
            type: integer
            format: int64
          example: 1
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CardRequest'
        required: true
      responses:
        '200':
          description: Карта успешно создана
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/CardResponse'
        '400':
          description: Неверные данные карты
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Не авторизован
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Доступ запрещен
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Пользователь не найден
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /admin/users:
    get:
      tags:
        - Admin Controller
      summary: Получить всех пользователей
      description: >-
        Возвращает список всех пользователей с пагинацией и поиском. Только для
        ADMIN
      operationId: getAllUsers
      parameters:
        - name: search
          in: query
          description: Поиск по username или email (частичное совпадение)
          required: false
          schema:
            type: string
        - name: pageable
          in: query
          description: >-
            Пагинация: page, size, sort. Пример:
            page=0&size=10&sort=createdAt,desc
          required: true
          schema:
            $ref: '#/components/schemas/Pageable'
      responses:
        '200':
          description: Список пользователей успешно получен
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/PageUserResponse'
        '400':
          description: Неверные параметры запроса
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Не авторизован
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Доступ запрещен - требуется роль ADMIN
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /admin/users/{userId}:
    get:
      tags:
        - Admin Controller
      summary: Получить пользователя по ID
      description: Возвращает детальную информацию о пользователе. Только для ADMIN
      operationId: getUserById
      parameters:
        - name: userId
          in: path
          description: ID пользователя
          required: true
          schema:
            type: integer
            format: int64
          example: 1
      responses:
        '200':
          description: Пользователь найден
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/UserResponse'
        '400':
          description: Неверный ID пользователя
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Не авторизован
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Доступ запрещен
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Пользователь не найден
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      tags:
        - Admin Controller
      summary: Удалить пользователя
      description: >-
        Удаляет пользователя из системы. Все карты пользователя также будут
        удалены. Только для ADMIN
      operationId: deleteUser
      parameters:
        - name: userId
          in: path
          description: ID пользователя
          required: true
          schema:
            type: integer
            format: int64
          example: 1
      responses:
        '204':
          description: Пользователь успешно удален
        '400':
          description: Неверный ID пользователя
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Не авторизован
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Доступ запрещен
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Пользователь не найден
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /admin/cards:
    get:
      tags:
        - Admin Controller
      summary: Получить все карты
      description: >-
        Возвращает список всех карт в системе с пагинацией и поиском. Только для
        ADMIN
      operationId: getAllCards
      parameters:
        - name: search
          in: query
          description: 'Поиск по маске номера карты (например: 1234)'
          required: false
          schema:
            type: string
        - name: pageable
          in: query
          description: 'Пагинация: page, size, sort'
          required: true
          schema:
            $ref: '#/components/schemas/Pageable'
      responses:
        '200':
          description: Список карт успешно получен
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/PageCardResponse'
        '400':
          description: Неверные параметры запроса
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Не авторизован
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Доступ запрещен - требуется роль ADMIN
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /admin/cards/{cardId}:
    delete:
      tags:
        - Admin Controller
      summary: Удалить карту
      description: Удаляет карту из системы. Только для ADMIN
      operationId: deleteCard
      parameters:
        - name: cardId
          in: path
          description: ID карты
          required: true
          schema:
            type: integer
            format: int64
          example: 1
      responses:
        '204':
          description: Карта успешно удалена
        '400':
          description: Неверный ID карты
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Не авторизован
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Доступ запрещен
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Карта не найдена
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorResponse'
components:
  schemas:
    ErrorResponse:
      type: object
      properties:
        timestamp:
          type: string
          description: Время возникновения ошибки
          format: date-time
        status:
          type: integer
          description: HTTP статус код
          format: int32
          example: 400
        error:
          type: string
          description: Тип ошибки
          example: Bad Request
        message:
          type: string
          description: Сообщение об ошибке
          example: 'User not found with id: 1'
        path:
          type: string
          description: Путь запроса
          example: /api/users/1
        validationErrors:
          type: object
          additionalProperties:
            type: string
            description: Детальные ошибки валидации (поле -> сообщение)
            example: '{"username":"Username is required","email":"Invalid email format"}'
          description: Детальные ошибки валидации (поле -> сообщение)
          example:
            username: Username is required
            email: Invalid email format
      description: Ответ с информацией об ошибке
    UpdateRoleRequest:
      required:
        - role
      type: object
      properties:
        role:
          type: string
          description: Новая роль пользователя
          example: ADMIN
          enum:
            - USER
            - ADMIN
      description: Новая роль
    UserResponse:
      type: object
      properties:
        id:
          type: integer
          description: ID пользователя
          format: int64
          example: 1
        username:
          type: string
          description: Имя пользователя (уникальное)
          example: john_doe
        email:
          type: string
          description: Email пользователя (уникальный)
          example: john@example.com
        role:
          type: string
          description: Роль пользователя в системе
          example: USER
          enum:
            - USER
            - ADMIN
        createdAt:
          type: string
          description: Дата и время регистрации
          format: date-time
      description: Ответ с данными пользователя
    CardResponse:
      type: object
      properties:
        id:
          type: integer
          description: ID карты
          format: int64
          example: 1
        cardNumberMasked:
          type: string
          description: Маскированный номер карты
          example: '**** **** **** 1234'
        cardHolderName:
          type: string
          description: Имя владельца на карте
          example: IVAN PETROV
        expirationDate:
          type: string
          description: Срок действия
          format: date
          example: '2025-12-31'
        status:
          type: string
          description: Статус карты
          example: ACTIVE
          enum:
            - ACTIVE
            - BLOCKED
            - EXPIRED
        balance:
          type: number
          description: Баланс
          example: 15000.5
        userId:
          type: integer
          description: ID владельца карты
          format: int64
          example: 1
      description: Ответ с данными банковской карты
...
    CardRequest:
      required:
        - cardHolderName
        - expirationDate
      type: object
      properties:
        cardHolderName:
          maxLength: 100
          minLength: 2
          type: string
        expirationDate:
          type: string
          format: date
      description: Данные новой карты
...
    Pageable:
      type: object
      properties:
        page:
          minimum: 0
          type: integer
          format: int32
        size:
          minimum: 1
          type: integer
          format: int32
        sort:
          type: array
          items:
            type: string
    PageCardResponse:
      type: object
      properties:
        content:
          type: array
          description: Список карт на текущей странице
          items:
            $ref: '#/components/schemas/CardResponse'
        number:
          type: integer
          description: Номер текущей страницы (начиная с 0)
          format: int32
          example: 0
        size:
          type: integer
          description: Размер страницы
          format: int32
          example: 10
        totalElements:
          type: integer
          description: Общее количество элементов
          format: int64
          example: 25
        totalPages:
          type: integer
          description: Общее количество страниц
          format: int32
          example: 3
        hasNext:
          type: boolean
          description: Есть ли следующая страница
          example: true
        hasPrevious:
          type: boolean
          description: Есть ли предыдущая страница
          example: false
        first:
          type: boolean
          description: Первая ли это страница
          example: true
        last:
          type: boolean
          description: Последняя ли это страница
          example: false
      description: Страница с картами
...
  securitySchemes:
    bearerAuth:
      type: http
      name: Authorization
      in: header
      scheme: bearer
      bearerFormat: JWT